refactor(web): formalize machine budget modules
