+++ web/src/lib/features/machines/components/machine-create-wizard.svelte
@@ -1,26 +1,29 @@
 <script lang="ts">
-  import * as Dialog from '$ui/dialog'
-  import { cn } from '$lib/utils'
+  import type { Machine, MachineSSHBootstrapResult } from '$lib/api/contracts'
   import { i18nStore } from '$lib/i18n/store.svelte'
   import { toastStore } from '$lib/stores/toast.svelte'
-  import type { Machine, MachineSSHBootstrapResult } from '$lib/api/contracts'
+  import * as Dialog from '$ui/dialog'
   import MachineCreateWizardFooter from './machine-create-wizard-footer.svelte'
+  import MachineCreateWizardProgress from './machine-create-wizard-progress.svelte'
+  import {
+    applyMachineWizardLocationChoice,
+    canAdvanceMachineWizardStep,
+    computeMachineWizardStepOrder,
+    createMachineCreateWizardDraft,
+  } from './machine-create-wizard-flow'
   import {
     machineWizardLocationOptions,
     machineWizardStrategyOptions,
   } from './machine-create-wizard-options'
   import MachineCreateWizardReview from './machine-create-wizard-review.svelte'
   import {
-    machineErrorMessage,
-    runMachineHealthRefresh,
-    runMachineSSHBootstrap,
-    saveMachine,
-  } from './machines-page-api'
+    MachineCreateWizardSubmitError,
+    submitMachineCreateWizard,
+  } from './machine-create-wizard-submit'
   import MachineCreateWizardSteps from './machine-create-wizard-steps.svelte'
-  import type { MachineMutationInput } from '../types'
+  import { machineErrorMessage } from './machines-page-api'
   import type {
     MachineWizardLocationAnswer,
-    MachineWizardStep,
     MachineWizardStrategy,
   } from './machine-create-wizard-types'
 
@@ -34,166 +37,53 @@
     onCreated?: (machine: Machine) => void
   } = $props()
 
-  let step = $state<MachineWizardStep>('location')
-  let location = $state<MachineWizardLocationAnswer | null>(null)
-  let name = $state('')
-  let host = $state('')
-  let strategy = $state<MachineWizardStrategy | null>(null)
-  let sshUser = $state('')
-  let sshKeyPath = $state('~/.ssh/id_ed25519')
-  let advertisedEndpoint = $state('')
-
+  let wizard = $state(createMachineCreateWizardDraft())
   let saving = $state(false)
   let bootstrapping = $state(false)
   let bootstrapResult = $state<MachineSSHBootstrapResult | null>(null)
   let errorMessage = $state('')
 
   $effect(() => {
-    if (open) {
-      step = 'location'
-      location = null
-      name = ''
-      host = ''
-      strategy = null
-      sshUser = ''
-      sshKeyPath = '~/.ssh/id_ed25519'
-      advertisedEndpoint = ''
-      saving = false
-      bootstrapping = false
-      bootstrapResult = null
-      errorMessage = ''
-    }
+    if (!open) return
+    wizard = createMachineCreateWizardDraft()
+    saving = false
+    bootstrapping = false
+    bootstrapResult = null
+    errorMessage = ''
   })
 
-  const strategyNeedsSSH = $derived(strategy === 'ssh-install-listener' || strategy === 'reverse')
-  const stepOrder = $derived(computeStepOrder(location, strategy))
-  const currentStepIndex = $derived(Math.max(0, stepOrder.indexOf(step)))
+  const strategyNeedsSSH = $derived(
+    wizard.strategy === 'ssh-install-listener' || wizard.strategy === 'reverse',
+  )
+  const stepOrder = $derived(computeMachineWizardStepOrder(wizard.location, wizard.strategy))
+  const currentStepIndex = $derived(Math.max(0, stepOrder.indexOf(wizard.step)))
   const totalSteps = $derived(stepOrder.length)
-  const isLastStep = $derived(step === 'review')
-
-  function computeStepOrder(
-    loc: MachineWizardLocationAnswer | null,
-    strat: MachineWizardStrategy | null,
-  ): MachineWizardStep[] {
-    if (loc === 'local') return ['location', 'identity', 'review']
-    const order: MachineWizardStep[] = ['location', 'identity', 'strategy']
-    if (strat === 'direct-open') order.push('advertised-endpoint')
-    if (strat === 'ssh-install-listener' || strat === 'reverse') {
-      order.push('credentials')
-    }
-    order.push('review')
-    return order
-  }
-
-  function canAdvance(): boolean {
-    switch (step) {
-      case 'location':
-        return location !== null
-      case 'identity':
-        if (location === 'local') return name.trim().length > 0
-        return name.trim().length > 0 && host.trim().length > 0
-      case 'strategy':
-        return strategy !== null
-      case 'credentials':
-        return sshUser.trim().length > 0 && sshKeyPath.trim().length > 0
-      case 'advertised-endpoint':
-        return advertisedEndpoint.trim().length > 0
-      case 'review':
-        return true
-    }
-  }
+  const isLastStep = $derived(wizard.step === 'review')
+  const canAdvanceCurrentStep = $derived(canAdvanceMachineWizardStep(wizard))
 
   function goNext() {
-    if (!canAdvance()) return
-    const idx = stepOrder.indexOf(step)
-    if (idx < stepOrder.length - 1) {
-      step = stepOrder[idx + 1]
+    if (!canAdvanceCurrentStep) return
+    const nextIndex = stepOrder.indexOf(wizard.step) + 1
+    if (nextIndex < stepOrder.length) {
+      wizard.step = stepOrder[nextIndex]
     }
   }
 
   function goBack() {
-    const idx = stepOrder.indexOf(step)
-    if (idx > 0) {
-      step = stepOrder[idx - 1]
+    const previousIndex = stepOrder.indexOf(wizard.step) - 1
+    if (previousIndex >= 0) {
+      wizard.step = stepOrder[previousIndex]
     }
   }
 
   function pickLocation(value: MachineWizardLocationAnswer) {
-    location = value
-    if (value === 'local') {
-      strategy = null
-      name = name || 'local'
-    } else if (!strategy) {
-      strategy = 'ssh-install-listener'
-    }
+    wizard = applyMachineWizardLocationChoice(wizard, value)
     setTimeout(() => goNext(), 120)
   }
 
   function pickStrategy(value: MachineWizardStrategy) {
-    strategy = value
+    wizard.strategy = value
     setTimeout(() => goNext(), 120)
-  }
-
-  function buildMutationInput(): MachineMutationInput {
-    if (location === 'local') {
-      return {
-        name: name.trim() || 'local',
-        host: 'local',
-        port: 22,
-        reachability_mode: 'local',
-        execution_mode: 'local_process',
-        ssh_user: '',
-        ssh_key_path: '',
-        advertised_endpoint: '',
-        description: '',
-        labels: [],
-        status: 'online',
-        workspace_root: '',
-        agent_cli_path: '',
-        env_vars: [],
-      }
-    }
-
-    const hostValue = host.trim()
-    if (strategy === 'reverse') {
-      return {
-        name: name.trim(),
-        host: hostValue,
-        port: 22,
-        reachability_mode: 'reverse_connect',
-        execution_mode: 'websocket',
-        ssh_user: sshUser.trim(),
-        ssh_key_path: sshKeyPath.trim(),
-        advertised_endpoint: '',
-        description: '',
-        labels: [],
-        status: 'offline',
-        workspace_root: '',
-        agent_cli_path: '',
-        env_vars: [],
-      }
-    }
-
-    const endpoint =
-      strategy === 'direct-open'
-        ? advertisedEndpoint.trim()
-        : `ws://${hostValue}:19837/openase/runtime`
-    return {
-      name: name.trim(),
-      host: hostValue,
-      port: 22,
-      reachability_mode: 'direct_connect',
-      execution_mode: 'websocket',
-      ssh_user: strategy === 'ssh-install-listener' ? sshUser.trim() : '',
-      ssh_key_path: strategy === 'ssh-install-listener' ? sshKeyPath.trim() : '',
-      advertised_endpoint: endpoint,
-      description: '',
-      labels: [],
-      status: 'offline',
-      workspace_root: '',
-      agent_cli_path: '',
-      env_vars: [],
-    }
   }
 
   async function handleCreate() {
@@ -204,53 +94,35 @@
 
     errorMessage = ''
     saving = true
-    let createdMachine: Machine | null = null
     try {
-      const mutation = buildMutationInput()
-      const created = await saveMachine(organizationId, null, 'create', mutation)
-      createdMachine = created.machine
-
-      if (strategy === 'ssh-install-listener' || strategy === 'reverse') {
-        bootstrapping = true
-        try {
-          const topology =
-            strategy === 'ssh-install-listener' ? 'remote-listener' : 'reverse-connect'
-          const listenerAddress = strategy === 'ssh-install-listener' ? '0.0.0.0:19837' : ''
-          bootstrapResult = await runMachineSSHBootstrap(created.machine.id, {
-            topology,
-            ...(listenerAddress ? { listener_address: listenerAddress } : {}),
-          })
-          const refreshed = await runMachineHealthRefresh(created.machine.id)
-          createdMachine = refreshed.machine
-          toastStore.success(
-            bootstrapResult.summary ||
-              i18nStore.t('machines.machineCreateWizard.successes.bootstrap'),
-          )
-        } catch (bootstrapError) {
-          errorMessage = machineErrorMessage(
-            bootstrapError,
-            i18nStore.t('machines.machineCreateWizard.errors.bootstrap'),
-          )
-          toastStore.error(errorMessage)
-        } finally {
-          bootstrapping = false
-        }
-      } else {
-        toastStore.success(i18nStore.t('machines.machineCreateWizard.successes.created'))
-      }
-
-      if (createdMachine) {
-        onCreated?.(createdMachine)
-      }
-
-      if (!bootstrapResult && !errorMessage) {
+      const created = await submitMachineCreateWizard({
+        organizationId,
+        draft: wizard,
+        setBootstrapping: (value) => (bootstrapping = value),
+      })
+      bootstrapResult = created.bootstrapResult
+      onCreated?.(created.machine)
+      toastStore.success(
+        created.bootstrapResult?.summary ||
+          i18nStore.t(
+            created.bootstrapResult
+              ? 'machines.machineCreateWizard.successes.bootstrap'
+              : 'machines.machineCreateWizard.successes.created',
+          ),
+      )
+      if (!created.bootstrapResult) {
         open = false
       }
     } catch (caughtError) {
-      errorMessage = machineErrorMessage(
-        caughtError,
-        i18nStore.t('machines.machineCreateWizard.errors.create'),
-      )
+      const stage =
+        caughtError instanceof MachineCreateWizardSubmitError ? caughtError.stage : 'create'
+      const fallbackKey =
+        stage === 'bootstrap'
+          ? 'machines.machineCreateWizard.errors.bootstrap'
+          : 'machines.machineCreateWizard.errors.create'
+      const cause =
+        caughtError instanceof MachineCreateWizardSubmitError ? caughtError.cause : caughtError
+      errorMessage = machineErrorMessage(cause, i18nStore.t(fallbackKey))
       toastStore.error(errorMessage)
     } finally {
       saving = false
@@ -271,31 +143,17 @@
       </Dialog.Description>
     </Dialog.Header>
 
-    <div class="flex items-center justify-between gap-1 px-1 pt-1">
-      <div class="flex flex-1 gap-1">
-        {#each Array(totalSteps) as _, idx (idx)}
-          <span
-            class={cn(
-              'h-1 flex-1 rounded-full transition-colors',
-              idx <= currentStepIndex ? 'bg-primary' : 'bg-muted',
-            )}
-          ></span>
-        {/each}
-      </div>
-      <span class="text-muted-foreground shrink-0 text-[10px] font-medium">
-        {currentStepIndex + 1} / {totalSteps}
-      </span>
-    </div>
+    <MachineCreateWizardProgress {currentStepIndex} {totalSteps} />
 
     <Dialog.Body class="min-h-[220px] py-4">
-      {#if step === 'review'}
+      {#if wizard.step === 'review'}
         <MachineCreateWizardReview
-          {location}
-          {name}
-          {host}
-          {strategy}
-          {sshUser}
-          {advertisedEndpoint}
+          location={wizard.location}
+          name={wizard.name}
+          host={wizard.host}
+          strategy={wizard.strategy}
+          sshUser={wizard.sshUser}
+          advertisedEndpoint={wizard.advertisedEndpoint}
           {strategyNeedsSSH}
           {saving}
           {bootstrapping}
@@ -305,14 +163,14 @@
         />
       {:else}
         <MachineCreateWizardSteps
-          {step}
-          {location}
-          bind:name
-          bind:host
-          {strategy}
-          bind:sshUser
-          bind:sshKeyPath
-          bind:advertisedEndpoint
+          step={wizard.step}
+          location={wizard.location}
+          bind:name={wizard.name}
+          bind:host={wizard.host}
+          strategy={wizard.strategy}
+          bind:sshUser={wizard.sshUser}
+          bind:sshKeyPath={wizard.sshKeyPath}
+          bind:advertisedEndpoint={wizard.advertisedEndpoint}
           locationOptions={machineWizardLocationOptions}
           strategyOptions={machineWizardStrategyOptions}
           onPickLocation={pickLocation}
@@ -324,7 +182,7 @@
     <Dialog.Footer>
       <MachineCreateWizardFooter
         {currentStepIndex}
-        canAdvance={canAdvance()}
+        canAdvance={canAdvanceCurrentStep}
         {isLastStep}
         {saving}
         {bootstrapping}
